+++ Changelog.md
@@ -1,4 +1,10 @@
-# next
+# 0.3
+
+- Rename `OneTuple` to `Solo`
+- Add `Typeable`, `Data`, `MonadZip`, `Eq1`, `Ord1`, `Show1`, `Read1`,
+  `Generic` and `Generic1` instances
+
+# 0.2.2.1
 
 - Compatible with GHC-8.6